pyproject.toml
[tool.poetry]
name = "personal-chatbot"
version = "0.0.1"
description = "Personal AI chatbot based on Langchain linked with various LLM models"
authors = ["Oleg Mikhailovskii"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
nltk = "^3.8.1"
python-dotenv = "^1.0.1"
langchain-anthropic = "^0.1.8"
langchain-cohere = "^0.1.2"
langchain-google-genai = "^1.0.2"
langchain-groq = "^0.1.2"
langchain = "^0.1.16"
langchain-community = "^0.0.34"

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
coverage = "^7.4.4"
flake8 = "^7.0.0"
pre-commit = "^3.7.0"
pytest = "^8.1.1"
ruff = "^0.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
